cleanup and refactor (before BIG REFACTOR)
