+++ lib/npc-input.js
@@ -1,7 +1,7 @@
 'use strict'
 
 // import { GURPS } from '../module/gurps.js'
-import { GurpsActor, Advantage, Skill, Melee, Ranged, Encumbrance, Note, Equipment } from '../module/actor.js'
+import { GurpsActor, Advantage, Skill, Melee, Ranged, Encumbrance, Note, Equipment } from '../module/actor/actor.js'
 import { digitsOnly } from './jquery-helper.js'
 import { DamageTables } from '../module/damage/damage-tables.js'
 import { i18n, sanitize } from './utilities.js'
@@ -30,7 +30,7 @@ Hooks.on(`renderNpcInput`, (app, html, data) => {
   $(html).find('#npc-input-name').focus()
 })
 
-//Hooks.once("ready", () => { new NpcInput().render(true) });
+// Hooks.once("ready", () => { new NpcInput().render(true) });
 
 const COMMENT_CHAR = '#'
 const ERR = `???:`
@@ -128,7 +128,7 @@ export class NpcInput extends FormApplication {
       if (ev.type == 'click' || (ev.type == 'keydown' && ev.which == 13)) this.importStatBlock(ev)
     })
     this.createButton = html.find('#npc-input-create')[0]
- }
+  }
 
   async importStatBlock(ev) {
     ev.preventDefault()
@@ -773,8 +773,7 @@ export class NpcInput extends FormApplication {
     } else {
       attblk = this.nextToken('Traits:', 'Advantages/Disadvantages:') // Look for either as start of ads/disads
       if (!attblk) {
-        if (this.peek('Skills:'))
-          attblk = this.nextToken('Skills:', 'junk')
+        if (this.peek('Skills:')) attblk = this.nextToken('Skills:', 'junk')
         else if (!this.peek('\nWeapons:')) this.mook.melee = COMMENT_CHAR + 'No attacks found' // If Weapons aren't listedt later, show error.
       }
     }
