+++ module/item-sheet.js
@@ -1,5 +1,5 @@
 'use strict'
-import { Melee, Ranged, Skill, Spell, Advantage } from './actor.js'
+import { Melee, Ranged, Skill, Spell, Advantage } from './actor/actor.js'
 import { digitsAndDecimalOnly, digitsOnly } from '../lib/jquery-helper.js'
 import { recurselist } from '../lib/utilities.js'
 
@@ -26,7 +26,7 @@ export class GurpsItemSheet extends ItemSheet {
     sheetData.data.eqt.f_count = this.item.data.data.eqt.count // hack for Furnace module
     sheetData.name = this.item.name
     if (!this.item.data.data.globalid && !this.item.parent)
-      this.item.update({ 'data.globalid': this.item.id, '_id': this.item.id })
+      this.item.update({ 'data.globalid': this.item.id, _id: this.item.id })
     return sheetData
   }
 
@@ -43,17 +43,17 @@ export class GurpsItemSheet extends ItemSheet {
       let nm = ev.currentTarget.value
       let commit = {
         'data.eqt.name': nm,
-        name: nm
+        name: nm,
       }
       recurselist(this.item.data.data.melee, (e, k, d) => {
-        commit = {...commit, ...{ ['data.melee.' + k + ".name"]: nm }}
+        commit = { ...commit, ...{ ['data.melee.' + k + '.name']: nm } }
       })
       recurselist(this.item.data.data.ranged, (e, k, d) => {
-        commit = {...commit, ...{ ['data.melee.' + k + ".name"]: nm }}
+        commit = { ...commit, ...{ ['data.melee.' + k + '.name']: nm } }
       })
       this.item.update(commit)
     })
-//    html.find('#quantity').change(ev => this.item.update({ 'data.eqt.count': parseInt(ev.currentTarget.value) }))
+    //    html.find('#quantity').change(ev => this.item.update({ 'data.eqt.count': parseInt(ev.currentTarget.value) }))
 
     html.find('#add-melee').click(ev => {
       ev.preventDefault()
@@ -90,10 +90,10 @@ export class GurpsItemSheet extends ItemSheet {
       let r = new Advantage()
       this._addToList('ads', r)
     })
-    
+
     html.find('textarea').on('drop', this.dropFoundryLinks)
     html.find('input').on('drop', this.dropFoundryLinks)
-    
+
     html.find('.itemdraggable').each((_, li) => {
       li.setAttribute('draggable', true)
       li.addEventListener('dragstart', ev => {
@@ -109,14 +109,13 @@ export class GurpsItemSheet extends ItemSheet {
             type: 'Item',
             id: this.item.id,
             pack: this.item.pack,
-            data: this.item.data
+            data: this.item.data,
           })
         )
       })
     })
-
   }
-  
+
   dropFoundryLinks(ev) {
     if (!!ev.originalEvent) ev = ev.originalEvent
     let dragData = JSON.parse(ev.dataTransfer.getData('text/plain'))
@@ -134,13 +133,10 @@ export class GurpsItemSheet extends ItemSheet {
       n = game.items.get(dragData.id).name
     }
     if (!!n) {
-      let add = ` [${dragData.type}[${dragData.id}]` + '{' + n + '}]'    
+      let add = ` [${dragData.type}[${dragData.id}]` + '{' + n + '}]'
       $(ev.currentTarget).val($(ev.currentTarget).val() + add)
     }
- }
-
-
-
+  }
 
   async _deleteKey(ev) {
     let key = ev.currentTarget.getAttribute('name')
@@ -173,7 +169,7 @@ export class GurpsItemSheet extends ItemSheet {
 
   close() {
     super.close()
-    this.item.update({"data.eqt.name": this.item.name})
+    this.item.update({ 'data.eqt.name': this.item.name })
     if (!!this.object.editingActor) this.object.editingActor.updateItem(this.object)
   }
 }
